+++ tt_metal/sources.cmake
@@ -27,7 +27,6 @@ set(TT_METAL_PUBLIC_API
     api/tt-metalium/experimental/core_subset_write/buffer_write.hpp
     api/tt-metalium/experimental/core_subset_write/mesh_command_queue.hpp
     api/tt-metalium/experimental/core_subset_write/tensor.hpp
-    api/tt-metalium/experimental/dataflow_buffer/dataflow_buffer.hpp
     api/tt-metalium/experimental/device.hpp
     api/tt-metalium/experimental/disaggregation/kv_chunk_address_table.hpp
     api/tt-metalium/experimental/dispatch_context.hpp
@@ -50,7 +49,6 @@ set(TT_METAL_PUBLIC_API
     api/tt-metalium/experimental/fabric/topology_solver.hpp
     api/tt-metalium/experimental/fabric/topology_solver.tpp
     api/tt-metalium/experimental/forge_backdoor/global_semaphore.hpp
-    api/tt-metalium/experimental/host_api.hpp
     api/tt-metalium/experimental/inspector.hpp
     api/tt-metalium/experimental/inspector_config.hpp
     api/tt-metalium/experimental/kernel_cache.hpp
